+++ docs/releases.md
@@ -4,6 +4,12 @@ uid: releases
 
 # Releases
 
+## [4.6.2](https://github.com/BEagle1984/silverback/releases/tag/v4.6.2)
+
+### Fixes
+
+* Fix distributed background services graceful shutdown [[#247](https://github.com/BEagle1984/silverback/pull/247)]
+
 ## [4.6.1](https://github.com/BEagle1984/silverback/releases/tag/v4.6.1)
 
 ### Fixes